Clothing_Advisor_GUI/requirements.txt
tk
Pillow
configparser
requests
google-generativeai
